--- src/main/java/com/projectd/interpreter/semantic/symbol/table/RootStNode.java
@@ -1,8 +1,0 @@
-package com.projectd.interpreter.semantic.symbol.table;
-
-public class RootStNode extends StNode {
-
-    public RootStNode() {
-        super(null, null, null, StNodeScope.GLOBAL);
-    }
-}
